cv-models-service/requirements.txt
flask
ultralytics
torch
torchvision
opencv-python
numpy
pillow
boto3
python-dotenv